+++ app/src/main/java/com/exam/shops/WeekEntry.java
@@ -2,9 +2,9 @@ package com.exam.shops;
 public class WeekEntry {
     private final String weekTitle;
     private final String dateRange;
-    private final int amount;
+    private final float amount;
 
-    public WeekEntry(String weekTitle, String dateRange, int amount) {
+    public WeekEntry(String weekTitle, String dateRange, float amount) {
         this.weekTitle = weekTitle;
         this.dateRange = dateRange;
         this.amount = amount;
@@ -12,6 +12,6 @@ public class WeekEntry {
 
     public String getWeekTitle() { return weekTitle; }
     public String getDateRange() { return dateRange; }
-    public int getAmount() { return amount; }
+    public float getAmount() { return amount; }
 }
